+++ monoloco/eval/__init__.py
@@ -2,4 +2,3 @@
 from .eval_kitti import EvalKitti
 from .generate_kitti import GenerateKitti
 from .geom_baseline import geometric_baseline
-
