+++ monoloco/network/architectures.py
