+++ monoloco/network/net.py
@@ -8,9 +8,8 @@ from collections import defaultdict
 
 import torch
 
-from ..utils import get_iou_matches, reorder_matches, get_keypoints, pixel_to_camera, xyz_from_distance, \
-    get_monoloco_inputs, unnormalize_bi, laplace_sampling
-from ..train import LinearModel
+from ..utils import get_iou_matches, reorder_matches, get_keypoints, pixel_to_camera, xyz_from_distance
+from . import LinearModel, preprocess_monoloco, unnormalize_bi, laplace_sampling
 
 
 class MonoLoco:
@@ -42,7 +41,7 @@ class MonoLoco:
             return None, None
 
         with torch.no_grad():
-            inputs = get_monoloco_inputs(torch.tensor(keypoints).to(self.device), torch.tensor(kk).to(self.device))
+            inputs = preprocess_monoloco(torch.tensor(keypoints).to(self.device), torch.tensor(kk).to(self.device))
             if self.n_dropout > 0:
                 self.model.dropout.training = True  # Manually reactivate dropout in eval
                 total_outputs = torch.empty((0, inputs.size()[0])).to(self.device)
