--- monoloco/predict/__init__.py
@@ -1,3 +1,0 @@
-
-from .pifpaf import PifPaf, ImageList
-from .network import MonoLoco
