+++ monoloco/run.py
@@ -102,7 +102,7 @@ def main():
             from .visuals.webcam import webcam
             webcam(args)
         else:
-            from .predict.predict import predict
+            from predict import predict
             predict(args)
 
     elif args.command == 'prep':
