+++ monoloco/train/__init__.py
@@ -1,4 +1,2 @@
 
-from .trainer import Trainer
 from .hyp_tuning import HypTuning
-from .architectures import LinearModel
