+++ monoloco/utils/__init__.py
@@ -1,10 +1,8 @@
 
-from .iou import get_iou_matches, reorder_matches
-from .misc import get_task_error, get_pixel_error
+from .iou import get_iou_matches, reorder_matches, get_iou_matrix
+from .misc import get_task_error, get_pixel_error, append_cluster
 from .kitti import check_conditions, get_category, split_training, parse_ground_truth, get_calibration
-from .pifpaf import preprocess_pif
-from .camera import xyz_from_distance, get_keypoints, pixel_to_camera, xyz_from_distance
-from .stereo import depth_from_disparity
-from .network import get_monoloco_inputs, unnormalize_bi, laplace_sampling
+from .camera import xyz_from_distance, get_keypoints, pixel_to_camera, project_3d
 from .logs import set_logger
-from .misc import get_task_error
+from .stereo import depth_from_disparity
+from ..utils.nuscenes import select_categories
